+++ Sources/MacIMEKit/Response.swift
@@ -1,0 +1,7 @@
+public enum ResponseStatus {
+   case ok, err
+}
+public struct Response {
+   public var status: ResponseStatus
+   public var content: String
+}